Added key gen functionality to lando pull lagoon recipe.
